+++ tsconfig.json
@@ -7,7 +7,8 @@
 		],
 		"sourceMap": true,
 		"rootDir": ".",
-		"strict": true
+		"strict": true,
+		"typeRoots": ["./node_modules/@types"]
 	},
 	"include": [
 		"src",
